Rename the `x86_64` module to `x64`.
The latter is widely understood and shorter! This commit does the bare
minimum possible in terms of renaming.
