+++ ykrt/src/compile/jitc_yk/codegen/mod.rs
@@ -12,7 +12,7 @@ mod abs_stack;
 pub(crate) mod reg_alloc;
 
 #[cfg(target_arch = "x86_64")]
-pub(crate) mod x86_64;
+pub(crate) mod x64;
 
 /// A code generator.
 ///
@@ -29,7 +29,7 @@ pub(crate) trait CodeGen: Send + Sync {
 
 pub(crate) fn default_codegen() -> Result<Arc<dyn CodeGen>, Box<dyn Error>> {
     #[cfg(target_arch = "x86_64")]
-    return Ok(x86_64::X86_64CodeGen::new()?);
+    return Ok(x64::X86_64CodeGen::new()?);
 
     #[cfg(not(target_arch = "x86_64"))]
     return Err("No code generator available for this platform".into());
