+++ ykrt/src/compile/jitc_yk/codegen/x64/deopt.rs
